fix: satisfy social operation quality gates
